Split out QUIC specific network events
